Dans soaked smokes redux (#36979)
* Meat and Potatoes

* Whoops

Whoops

* entityTables my beloathed

reworked the cig fill for Dan's soaked smokes to use an entity table. There was just, a useless bit of code calling for a cig that didn't exist the in mixed cig parent for dan's and it was what was fucking it up and making it SO HARD for me to get it to work. Thanks Dan.

* ehhh, its good for the diff I guess

* Augh

* Why mappers why

* Welp

* migration.yml is pretty cool

Also like there is no reason for nasty and medical, they probably originally were made as a workaround.

(cherry picked from commit b589c75c2a4792eb6bb0b808a5ecaf72e751802c)